# Conflicts:
#	src/main/java/com/example/backend/common/response/ApiResponse.java
#	src/main/java/com/example/backend/common/response/ApiResponseFactory.java
#	src/main/java/com/example/backend/controller/PropertyController.java
#	src/main/java/com/example/backend/entity/Property.java
#	src/main/java/com/example/backend/repository/FavoriteRepository.java
#	src/main/java/com/example/backend/repository/PropertyRepository.java
#	src/main/java/com/example/backend/service/PropertyService.java

+++ src/main/java/com/example/backend/repository/FavoriteRepository.java
@@ -3,5 +3,7 @@ package com.example.backend.repository;
 import com.example.backend.entity.Favorite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FavoriteRepository extends JpaRepository<Favorite,Integer> {
+import com.example.backend.entity.Favorite;
+
+public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
 }
